+++ src/tests/utils.ts
@@ -1,17 +1,3 @@
-import { Action } from "../reducer";
-import {
-  parseBankTrade,
-  parseDiscardedMessage,
-  parseGot,
-  parseMonoplyCard,
-  parsePlayersTrade,
-  parsePurchase,
-  parsePurposalMessage,
-  parseStoleFromYouMessage,
-  parseStoleUnknownMessage,
-  parseYearofPlenty,
-  recognizeUsers,
-} from "../scripts/actionParser";
 import {
   PurchaseType,
   ResourceType,
@@ -20,8 +6,8 @@ import {
   Users,
 } from "../types";
 import { getImg } from "../utils/index.";
-import keywords from "../utils/keywords";
-import testData from "./data";
+
+import testData from "../utils/data";
 
 type UserProperties = {
   resources: UserResources;
@@ -29,43 +15,8 @@ type UserProperties = {
 };
 
 /**
- * Returns a random number between min (inclusive) and max (exclusive)
+ * Different variations on color and starting resources for players
  */
-export const getRandomArbitrary = (min: number, max: number) => {
-  return Math.floor(Math.random() * (max - min) + min);
-};
-
-export const shuffleArray = <T>(array: Array<T>) => {
-  let currentIndex = array.length,
-    randomIndex;
-
-  // While there remain elements to shuffle...
-  while (currentIndex != 0) {
-    // Pick a remaining element...
-    randomIndex = Math.floor(Math.random() * currentIndex);
-    currentIndex--;
-
-    // And swap it with the current element.
-    [array[currentIndex], array[randomIndex]] = [
-      array[randomIndex],
-      array[currentIndex],
-    ];
-  }
-
-  return array;
-};
-
-export const getRandomResources = (amount: number = 9) => {
-  const gainedResources: ResourceType[] = [];
-  const resourceAmount = getRandomArbitrary(1, amount);
-  for (let i = 0; i < resourceAmount; i++) {
-    const randomIndex = getRandomArbitrary(0, testData.resources.length);
-    gainedResources.push(testData.resources[randomIndex]);
-  }
-
-  return gainedResources;
-};
-
 const startingResourcesAndConfig: UserProperties[] = [
   {
     resources: {
@@ -109,6 +60,51 @@ const startingResourcesAndConfig: UserProperties[] = [
   },
 ];
 
+/**
+ * Returns a random number between min (inclusive) and max (exclusive)
+ */
+const getRandomArbitrary = (min: number, max: number) => {
+  return Math.floor(Math.random() * (max - min) + min);
+};
+
+/**
+ * Suffles array
+ * TODO: Can be refactored
+ */
+export const shuffleArray = <T>(array: Array<T>) => {
+  let currentIndex = array.length,
+    randomIndex;
+
+  // While there remain elements to shuffle...
+  while (currentIndex != 0) {
+    // Pick a remaining element...
+    randomIndex = Math.floor(Math.random() * currentIndex);
+    currentIndex--;
+
+    // And swap it with the current element.
+    [array[currentIndex], array[randomIndex]] = [
+      array[randomIndex],
+      array[currentIndex],
+    ];
+  }
+
+  return array;
+};
+
+export const getRandomResources = (amount: number = 9) => {
+  const gainedResources: ResourceType[] = [];
+  const resourceAmount = getRandomArbitrary(1, amount);
+  for (let i = 0; i < resourceAmount; i++) {
+    const randomIndex = getRandomArbitrary(0, testData.resources.length);
+    gainedResources.push(testData.resources[randomIndex]);
+  }
+
+  return gainedResources;
+};
+
+/**
+ * Reset all resources to zero
+ */
 export const emptyResources = {
   [ResourceType.WOOD]: 0,
   [ResourceType.WHEAT]: 0,
@@ -117,6 +113,11 @@ export const emptyResources = {
   [ResourceType.STONE]: 0,
 };
 
+/**
+ * Create Player with config and the starting resources
+ * @param allResourceStartEmpty if we want to assign emptry resource to all players
+ * @returns
+ */
 export const createPlayersAndProperties = (
   allResourceStartEmpty: boolean = false
 ) => {
@@ -124,7 +125,9 @@ export const createPlayersAndProperties = (
     startingResourcesAndConfig
   );
   return testData.users.reduce<Users>((acc, user, index) => {
+    // Assign shuffle amount of resource to player
     acc[user] = shuffledStartingResourcesAndConfig[index];
+    // Remove all resources from player if allResourceStartEmpty is true
     if (allResourceStartEmpty) {
       acc[user].resources = emptyResources;
     }
@@ -132,6 +135,14 @@ export const createPlayersAndProperties = (
   }, {});
 };
 
+/**
+ * Simulates the output of each log
+ * @param color
+ * @param user1
+ * @param keywords
+ * @param user2
+ * @returns
+ */
 export const createDivElement = (
   color: string,
   user1: string,
@@ -144,6 +155,12 @@ export const createDivElement = (
   return node;
 };
 
+/**
+ * Creates the card image that we see in the game log
+ * @param node
+ * @param imageType
+ * @returns
+ */
 export const createChildImgElement = (
   node: HTMLElement,
   imageType: ResourceType | PurchaseType
@@ -154,263 +171,18 @@ export const createChildImgElement = (
   return node;
 };
 
-export const initiateTestingPlayers = (
-  dipatch: React.Dispatch<Action>,
-  allResourceStartEmpty: boolean = false
-) => {
-  const users = createPlayersAndProperties(allResourceStartEmpty);
-
-  (Object.keys(users) as Array<keyof typeof users>).forEach((user) => {
-    const node = createDivElement(
-      users[user].config.color,
-      user,
-      keywords.placeInitialSettlementSnippet
-    );
-
-    (Object.keys(users[user].resources) as Array<ResourceType>).forEach(
-      (resource) => {
-        if (users[user].resources[resource] > 0)
-          for (
-            let index = 0;
-            index < users[user].resources[resource];
-            index++
-          ) {
-            createChildImgElement(node, resource);
-          }
-      }
-    );
-
-    recognizeUsers(node, dipatch);
-  });
-};
-
-export const giveResourcesToPlayer = (
-  dipatch: React.Dispatch<Action>,
-  user: string,
-  addedResouces: ResourceType[],
-  color: string
-) => {
-  const node = createDivElement(color, user, keywords.receivedResourcesSnippet);
-  addedResouces.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  parseGot(node, dipatch);
-};
-
 /**
- * Method to give players resources
- * @param dipatch
+ * Counts the total amount of a specific type that is found in an array
+ * @param arrayWithType
+ * @param typeToBeCounted
+ * @returns
  */
-export const givePlayersInitialResources = (
-  dispatch: React.Dispatch<Action>,
-  playersResources: Record<string, ResourceType[]>,
-  players: Users
+export const countTotalTypeInArray = <T>(
+  arrayWithType: T[],
+  typeToBeCounted: T
 ) => {
-  Object.entries(playersResources).forEach(([player, resources]) => {
-    giveResourcesToPlayer(
-      dispatch,
-      player,
-      resources,
-      players[player].config.color
-    );
-  });
-};
-
-export const maxMonopolyGain = (
-  resourcesInPlay: ResourceType[],
-  resourceToCount: ResourceType
-) => {
-  return resourcesInPlay.reduce(
-    (acc, resource) => (resourceToCount === resource ? acc + 1 : acc),
+  return arrayWithType.reduce(
+    (acc, type) => (typeToBeCounted === type ? acc + 1 : acc),
     0
   );
-};
-
-export const playerMakesPurchase = (
-  dipatch: React.Dispatch<Action>,
-  user: string,
-  purchaseType: PurchaseType,
-  color: string
-) => {
-  const node = createDivElement(
-    color,
-    user,
-    purchaseType === PurchaseType.DEVELOPMENT
-      ? keywords.boughtSnippet
-      : keywords.builtSnippet
-  );
-
-  createChildImgElement(node, purchaseType);
-
-  parsePurchase(node, dipatch);
-};
-
-export const bankTrade = (
-  dispatch: React.Dispatch<Action>,
-  user: string,
-  gave: ResourceType[],
-  took: ResourceType[],
-  color: string
-) => {
-  const node = createDivElement(color, user, keywords.tradeBankGaveSnippet);
-
-  gave.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  const textNode = document.createTextNode(keywords.tradeBankTookSnippet);
-  node.appendChild(textNode);
-
-  took.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  parseBankTrade(node, dispatch);
-};
-export const offerPurposal = (
-  dispatch: React.Dispatch<Action>,
-  offeringPlayer: string,
-  offer: ResourceType[],
-  want: ResourceType[],
-  color: string
-) => {
-  const node = createDivElement(color, offeringPlayer, keywords.proposal);
-
-  offer.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  const textNodeEnd = document.createTextNode(keywords.wants);
-  node.appendChild(textNodeEnd);
-
-  want.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  parsePurposalMessage(node, dispatch);
-};
-export const playerTrade = (
-  dispatch: React.Dispatch<Action>,
-  offeringPlayer: string,
-  agreedPlayer: string,
-  gave: ResourceType[],
-  took: ResourceType[],
-  color: string
-) => {
-  offerPurposal(dispatch, offeringPlayer, gave, took, color);
-
-  const node = createDivElement(
-    color,
-    offeringPlayer,
-    keywords.tradedWithSnippet
-  );
-
-  gave.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  const textNodeEnd = document.createTextNode(`${keywords.tradeEnd} `);
-  node.appendChild(textNodeEnd);
-
-  took.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  const textNodeAgree = document.createTextNode(
-    `${keywords.tradeAgreePlayer} ${agreedPlayer}`
-  );
-  node.appendChild(textNodeAgree);
-
-  parsePlayersTrade(node, dispatch);
-};
-
-export const monopoly = (
-  dispatch: React.Dispatch<Action>,
-  player: string,
-  color: string,
-  monopolizedResource: ResourceType,
-  amountStolen: number
-) => {
-  const prevNode = createDivElement(color, player, keywords.stoleAllOfSnippet);
-
-  const node = createDivElement(color, player, keywords.monoplyStole);
-
-  createChildImgElement(node, monopolizedResource);
-
-  const textNodeAgree = document.createTextNode(` ${amountStolen}`);
-  node.appendChild(textNodeAgree);
-
-  parseMonoplyCard(node, prevNode, dispatch);
-};
-
-export const yearOfPlenty = (
-  dispatch: React.Dispatch<Action>,
-  player: string,
-  color: string,
-  pickedResources: ResourceType[]
-) => {
-  const node = createDivElement(color, player, keywords.yearOfPlenty);
-
-  pickedResources.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  parseYearofPlenty(node, dispatch);
-};
-
-export const discardCards = (
-  dispatch: React.Dispatch<Action>,
-  player: string,
-  color: string,
-  discardedResources: ResourceType[]
-) => {
-  const node = createDivElement(color, player, keywords.discardedSnippet);
-
-  discardedResources.forEach((resource) => {
-    createChildImgElement(node, resource);
-  });
-
-  parseDiscardedMessage(node, dispatch);
-};
-
-export const stoleFromOrByYou = (
-  dispatch: React.Dispatch<Action>,
-  player: string,
-  color: string,
-  stolenResource: ResourceType,
-  isYouTheStealer: boolean = false
-) => {
-  const node = createDivElement(
-    color,
-    isYouTheStealer ? "You" : player,
-    isYouTheStealer
-      ? keywords.youStoleFromSnippet
-      : keywords.stoleFromYouSnippet
-  );
-
-  createChildImgElement(node, stolenResource);
-
-  const textNodeAgree = document.createTextNode(
-    `${isYouTheStealer ? player : "you"}`
-  );
-  node.appendChild(textNodeAgree);
-
-  parseStoleFromYouMessage(node, dispatch);
-};
-
-export const unknownSteal = (
-  dispatch: React.Dispatch<Action>,
-  victim: string,
-  stealer: string,
-  color: string
-) => {
-  const node = createDivElement(
-    color,
-    stealer,
-    `${keywords.stoleFromSnippet}`,
-    victim
-  );
-
-  parseStoleUnknownMessage(node, dispatch);
 };
